BUGFIX: rounding and segfault fixes
 - RAM calculations now use `double` instead of `int` and `float`
 - fixed segfaults with `done` and `swap_pred_maps`
 - pulled `aln_ptrs` init outside of P/R calculation function

+++ src/globals.cpp
@@ -482,7 +482,7 @@ void Globals::parse_args(int argc, char ** argv) {
                 ERROR("Option '--max-ram' used without providing max RAM");
             }
             try {
-                this->max_ram = std::stof(argv[i++]);
+                this->max_ram = std::stod(argv[i++]);
             } catch (const std::exception & e) {
                 ERROR("Invalid max RAM provided");
             }
